[Refactor] move `documentNode`-related codes into `./src/editor/model/documentNode`
