+++ src/editor/model/documentNode/documentNode.ts
