+++ src/editor/model/documentNode/documentNodeProvider.ts
@@ -3,23 +3,23 @@ import { panic } from "src/base/common/utilities/panic";
 import { Mutable } from "src/base/common/utilities/type";
 import { TokenEnum, MarkEnum } from "src/editor/common/markdown";
 import { ProseNodeType, ProseMarkType } from "src/editor/common/proseMirror";
-import { DocumentNode, DocumentMark } from "src/editor/model/parser/documentNode";
-import { Codespan } from "src/editor/model/parser/mark/codespan";
-import { Emphasis } from "src/editor/model/parser/mark/emphasis";
-import { Link } from "src/editor/model/parser/mark/link";
-import { Strong } from "src/editor/model/parser/mark/strong";
-import { Blockquote } from "src/editor/model/parser/node/blockquote";
-import { CodeBlock } from "src/editor/model/parser/node/codeBlock";
-import { Escape } from "src/editor/model/parser/node/escape";
-import { Heading } from "src/editor/model/parser/node/heading";
-import { HorizontalRule } from "src/editor/model/parser/node/horizontalRule";
-import { HTML, InlineHTML } from "src/editor/model/parser/node/html";
-import { Image } from "src/editor/model/parser/node/image";
-import { LineBreak } from "src/editor/model/parser/node/lineBreak";
-import { List, ListItem } from "src/editor/model/parser/node/list";
-import { Paragraph } from "src/editor/model/parser/node/paragraph";
-import { Space } from "src/editor/model/parser/node/space";
-import { Text } from "src/editor/model/parser/node/text";
+import { DocumentNode, DocumentMark } from "src/editor/model/documentNode/documentNode";
+import { Codespan } from "src/editor/model/documentNode/mark/codespan";
+import { Emphasis } from "src/editor/model/documentNode/mark/emphasis";
+import { Link } from "src/editor/model/documentNode/mark/link";
+import { Strong } from "src/editor/model/documentNode/mark/strong";
+import { Blockquote } from "src/editor/model/documentNode/node/blockquote";
+import { CodeBlock } from "src/editor/model/documentNode/node/codeBlock";
+import { Escape } from "src/editor/model/documentNode/node/escape";
+import { Heading } from "src/editor/model/documentNode/node/heading";
+import { HorizontalRule } from "src/editor/model/documentNode/node/horizontalRule";
+import { HTML, InlineHTML } from "src/editor/model/documentNode/node/html";
+import { Image } from "src/editor/model/documentNode/node/image";
+import { LineBreak } from "src/editor/model/documentNode/node/lineBreak";
+import { List, ListItem } from "src/editor/model/documentNode/node/list";
+import { Paragraph } from "src/editor/model/documentNode/node/paragraph";
+import { Space } from "src/editor/model/documentNode/node/space";
+import { Text } from "src/editor/model/documentNode/node/text";
 import { EditorSchema, TOP_NODE_NAME } from "src/editor/model/schema";
 
 /**
