+++ src/editor/model/documentNode/mark/link.ts
@@ -1,7 +1,7 @@
 import { MarkEnum } from "src/editor/common/markdown";
 import { EditorTokens } from "src/editor/common/model";
 import { ProseMark, ProseMarkSpec, ProseNode } from "src/editor/common/proseMirror";
-import { DocumentMark, IParseTokenStatus } from "src/editor/model/parser/documentNode";
+import { DocumentMark, IParseTokenStatus } from "src/editor/model/documentNode/documentNode";
 import { IDocumentParseState } from "src/editor/model/parser/parser";
 import { IDocumentMarkSerializationOptions } from "src/editor/model/serializer/serializer";
